+++ ArpSrc/ArpCollections/ArpIndexedNode.cpp
@@ -10,15 +10,15 @@
  * <URL:http://www.angryredplanet.com/>.
  */
 
-#include <assert.h>
+#include <cassert>
 
 #ifndef ARPCOLLECTIONS_ARPINDEXEDNODE_H
 #include "ArpCollections/ArpIndexedNode.h"
 #endif
 
-#include <stdio.h>
+#include <cstdio>
 
-long	nodeCount = 0;
+int32	nodeCount = 0;
 
 
 ArpIndexedNode::ArpIndexedNode(void *dataArg)
